Use splits client config
