+++ package.json
@@ -41,7 +41,6 @@
     "@tanstack/react-table": "^8.11.8",
     "@types/mdx": "^2.0.11",
     "@vercel/analytics": "^1.2.2",
-    "alchemy-sdk": "^3.3.1",
     "axios": "^1.6.7",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
